Refactor some event handlers, to better separate the client-side ones
