+++ src/main/java/baubles/client/ClientEventHandler.java
@@ -1,0 +1,42 @@
+package baubles.client;
+
+import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraftforge.client.event.ModelRegistryEvent;
+import net.minecraftforge.client.model.ModelLoader;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
+import net.minecraftforge.fml.client.FMLClientHandler;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+import net.minecraftforge.fml.common.gameevent.TickEvent.Phase;
+import net.minecraftforge.fml.common.gameevent.TickEvent.PlayerTickEvent;
+import net.minecraftforge.fml.relauncher.Side;
+import baubles.api.BaubleType;
+import baubles.api.IBauble;
+import baubles.common.items.ItemRing;
+import baubles.common.network.PacketHandler;
+import baubles.common.network.PacketOpenBaublesInventory;
+
+public class ClientEventHandler
+{
+	@SubscribeEvent
+	public void registerItemModels(ModelRegistryEvent event) {
+		ModelLoader.setCustomModelResourceLocation(ItemRing.RING, 0, new ModelResourceLocation("baubles:ring", "inventory"));
+	}
+
+	@SubscribeEvent
+	public void playerTick(PlayerTickEvent event) {
+		if (event.side == Side.CLIENT && event.phase == Phase.START ) {
+			if (ClientProxy.KEY_BAUBLES.isPressed() && FMLClientHandler.instance().getClient().inGameHasFocus) {
+					PacketHandler.INSTANCE.sendToServer(new PacketOpenBaublesInventory(event.player));
+			}
+		}
+	}
+
+	@SubscribeEvent
+	public void tooltipEvent(ItemTooltipEvent event) {
+		if (!event.getItemStack().isEmpty() && event.getItemStack().getItem() instanceof IBauble) {
+			BaubleType bt = ((IBauble)event.getItemStack().getItem()).getBaubleType(event.getItemStack());
+			event.getToolTip().add(TextFormatting.GOLD + net.minecraft.client.resources.I18n.format("name." + bt));
+		}
+	}
+}
