+++ src/main/java/baubles/common/CommonProxy.java
@@ -3,13 +3,12 @@ package baubles.common;
 
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.network.IGuiHandler;
 import baubles.common.container.ContainerPlayerExpanded;
-import baubles.common.event.KeyHandler;
+import baubles.common.event.EventHandlerEntity;
 
 public class CommonProxy implements IGuiHandler {
-
-	public KeyHandler keyHandler;
 
 	@Override
 	public Object getClientGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
@@ -28,7 +27,9 @@ public class CommonProxy implements IGuiHandler {
 		return null;
 	}
 
-	public void registerKeyBindings() { }
+	public void registerEventHandlers() {
+		MinecraftForge.EVENT_BUS.register(new EventHandlerEntity());
+	}
 
 	public void init() { }
 }
